refactor: organize _utils as a package, isolate plotting backends
Replace the monolithic _utils.py with a _utils/ package whose submodules
group helpers by topic (stack, coordinates, timing, io, atlas, plotting).
Functions inside _utils now use public names; the leading underscore on
the package conveys "internal API" instead of decorating every callable.

Promote _build_atlas_cmap_and_norm out of atlas/_structures.py into
_utils/atlas.py so plotting can reach it without importing a private name
across module boundaries.

Split plotting/image.py: move plot_napari, draw_napari_labels,
labels_from_layer, and _attach_roi_labels_to_napari to a new
plotting/napari.py so each file targets a single backend. Add
plotting/_utils.py for the small set of helpers shared between matplotlib
and napari (coerce_complex_to_magnitude, sort_coords_for_plot), mirroring
the existing registration/_utils.py pattern.

Document the convention in AGENTS.md under a new Module Organization
section so the same rules apply to future contributions.
