+++ src/confusius/_utils/__init__.py
@@ -1,0 +1,6 @@
+"""Cross-module utilities for confusius.
+
+Each submodule groups one topic. Callers import the submodule they need
+directly (e.g. ``from confusius._utils.coordinates import get_coordinate_spacings``)
+rather than from the package root.
+"""
