+++ src/confusius/iq/process.py
@@ -7,7 +7,7 @@ import numpy as np
 import numpy.typing as npt
 import xarray as xr
 
-from confusius._utils import find_stack_level
+from confusius._utils.stack import find_stack_level
 from confusius.iq.clutter_filters import (
     clutter_filter_butterworth,
     clutter_filter_svd_from_cumulative_energy,
