+++ src/confusius/multipose/consolidate.py
@@ -11,7 +11,7 @@ import numpy as np
 import numpy.typing as npt
 import xarray as xr
 
-from confusius._utils import find_stack_level
+from confusius._utils.stack import find_stack_level
 from confusius.timing import convert_time_reference
